# Fixed test issue in Simple.Data performance test