--- a/Dapper.Tests/project.json
+++ b/Dapper.Tests/project.json
@@ -23,8 +23,7 @@
       "target": "project"
     },
     "xunit": "2.2.0-beta2-build3300",
-    "dotnet-test-xunit": "2.2.0-preview2-build1029",
-    "Belgrade.Sql.Client": "0.6.2"
+    "dotnet-test-xunit": "2.2.0-preview2-build1029"
   },
   "buildOptions": {
     "warningsAsErrors": true,
@@ -92,8 +91,7 @@
           "SOMA",
           "SIMPLEDATA",
           "SQLITE",
-          "XUNIT2",
-          "BELGRADE"
+          "XUNIT2"
         ]
       },
       "frameworkAssemblies": {
@@ -132,7 +130,7 @@
         "dnxcore50"
       ],
       "buildOptions": {
-        "define": [ "ASYNC", "COREFX", "XUNIT2", "SQLITE","BELGRADE" ]
+        "define": [ "ASYNC", "COREFX", "XUNIT2", "SQLITE" ]
       },
       "dependencies": {
         "Microsoft.NETCore.App": {