@NickCraver 2017-03-14
Merged in, thank you for finding and fixing!